Resolved #0000081.

+++ netgore/trunk/DemoGame.Server/Items/ItemTemplates.cs
@@ -29,7 +29,7 @@ namespace DemoGame.Server
             var itemTemplates = query.Execute();
             foreach (ItemTemplate it in itemTemplates)
             {
-                _itemTemplates[it.ID] = it;
+                _itemTemplates[(int)it.ID] = it;
 
                 if (log.IsInfoEnabled)
                     log.InfoFormat("Loaded ItemTemplate `{0}`", it);
